Add click event listener to all buttons
All buttons have had event listeners added using querySelectorAll
with a forEach loop on the NodeList returned. This was done to
reduce the amount of code needed to add the click event listener
to all buttons.

A data-value attribute was added to all buttons
that will have a numeric function to facilitate easier type
conversion later as mathematical operation functions begin to take
shape.

Minor UI tweak to text position of display div.