Create new storeOperationNumber()
The storeOperationNumber function is used to make sure a valid
sequence of numbers is being input to the calculator. 0 as the only
digit should not be allowed. Struggled with handling the logic for
this so it's messy right now - will need to look making it more
concise when my brain is working better again :).

In creating the storeOperationsNumber function, most of the code
was removed from the renderDisplay function to the new store function.
renderDisplay now will only do just that.