Create two functions to aid in populating display
storeCurrentNumber is used to store the user input as a string.
This function can also be used to check for early stage validation
like making sure the series of numbers is not too long.

populateDisplay is used purely to write the desired values to the
calculator display.

A new class was added to all numeric buttons, mostly to enable
a more targeted selection when using querySelectorAll to add the
click event listener.

Renamed the special-buttons and operator-buttons classes to make it
singular for consistency.